Remove misspelt file (move contents).
